Updated WorldManager to selectively load/display chunks.
WorldManager now requires you to consistently update it with
the camera or player position so that it can display only the
relevant chunks.  Note that you can pass null as a location to
the setViewerInfo function to load the whole map, and not need
to update after that at all.