Modify WorldChunk loading to not be dependant on ocean and effects.
Ocean and Effects do not affect game playability, and thus are not
neccessarily required for a map to be entered.  Disabled them to
reflect this.  Additionally fixed minor issue with missing callback.

Note: This patch is actually due to an issue with map effect loading
not always calling back when it should be, causing map loads to fail.