Switched to a reference system for server targetted GameObjects.
Internally the reference is held by the GameObject itself for GC
friendliness.  Additionally to this, we maintain proxy objects for
objects that are targetted which we do not have visibility for,
this permits us to share the position state of these objects across
many objects which reference them.  When an object goes out-of-view,
it is demoted back to a ProxyObject rather than being destroyed
entirely, and all references automatically point to the new object.
This 'automatic remap' also occurs when a ProxyObject is promoted to
a complete GameObject (by having it come in-view).